==> 0_creation_data.ipynb <==
# Creation_data

- Script a vocation de récupérer des données de fichiers HTML
- Auteur : Benjamin ALLEAU / benjamin.fraime@gmail.com

# Packages

from bs4 import BeautifulSoup
import pandas as pd
import re
import os

# Fonctions


# Charger le fichier HTML
def recuperateur_resultat(path_file):
    with open(path_file, "r", encoding="utf-8") as file:
        soup = BeautifulSoup(file, "html.parser")

    # Extraire les matchs
    games = []
    match_elements = soup.find_all("div", class_="Match__Container-sc-exmma4-0")

    for match in match_elements:
        court = match.find("div", class_="subtitle").text.strip().replace("Court ", "")
        teams = match.find_all("div", class_="Match__TeamContainer-sc-exmma4-3")[0]
        # Utiliser une regex pour extraire les noms des joueurs
        equipes = re.findall(r'>([\wÀ-ÿ\.\-]+\s[\wÀ-ÿ\.\-]+/[\wÀ-ÿ\.\-]+\s[\wÀ-ÿ\.\-]+)<', str(teams))

        if len(equipes) == 2:
            eq1_joueurs = equipes[0].split("/")
            eq2_joueurs = equipes[1].split("/")

        ## Recuperation des scores
        scores = match.find_all("input", class_="MatchGame__StyledInput-sc-1asd8lj-1")
        scores1 = str(scores[0])
        match = re.search(r'value="(\d+)"', scores1)
        resultat1 = match.group(1) if match else None
        scores2 = str(scores[1])
        match = re.search(r'value="(\d+)"', scores2)
        resultat2 = match.group(1) if match else None  
        games.append([court,eq1_joueurs[0].strip(),eq1_joueurs[1].strip(),eq2_joueurs[0].strip(),eq2_joueurs[1].strip(),resultat1,resultat2])
            

    df = pd.DataFrame(games, columns=["Num Court", "Eq1 Joueur 1", "Eq1 Joueur 2", "Eq2 Joueur 1", "Eq2 Joueur 2", "Score Eq1", "Score Eq2"])
    return df

def apply_function_to_html_files(directory, function):
    """
    Applique une fonction à tous les fichiers HTML dans un répertoire donné.

    Args:
        directory (str): Le chemin du répertoire contenant les fichiers HTML.
        function (callable): La fonction à appliquer aux fichiers HTML.
    """
    # Parcourir tous les fichiers dans le répertoire
    list_res = []
    i = 1
    for filename in os.listdir(directory):
        print(i)
        print(filename)
        # Vérifier si le fichier a l'extension .html
        if filename.endswith('.html'):
            # Construire le chemin complet du fichier
            file_path = os.path.join(directory, filename)
            # Appliquer la fonction au fichier
            res_temp = function(file_path)
            res_temp["week"] = i
            print(len(res_temp))
            list_res.append(res_temp)
            i += 1

    return list_res


# Traitement des données 

directory_path = 'data'

# Appliquer la fonction example_function à tous les fichiers HTML dans le répertoire
list_res = apply_function_to_html_files(directory_path, recuperateur_resultat)
df_resultat = pd.concat(list_res)


# Affichage des résultats

df_resultat.tail(5)

df_resultat = df_resultat[~df_resultat['Score Eq1'].isnull()]


df_resultat

df_resultat.to_pickle("data_preprocess/df_resultat.pkl")

df_resultat.to_csv("data_preprocess/df_resultat.csv",index=False)
