==> 1_dm_model.ipynb <==
# dm_model
- Script pour préparer les données à la modélisation et à l'affichage dans l'app streamlit
- Auteur : Benjamin ALLEAU / benjamin.fraime@gmail.com

# Package

import pandas as pd


# Charger le fichier pickle
df = pd.read_pickle('data_preprocess/df_resultat.pkl')

# Transformer le tableau en format détaillé pour chaque joueur
expanded_rows = []

for _, row in df.iterrows():
    num_court = row["Num Court"]
    eq1_j1, eq1_j2 = row["Eq1 Joueur 1"], row["Eq1 Joueur 2"]
    eq2_j1, eq2_j2 = row["Eq2 Joueur 1"], row["Eq2 Joueur 2"]
    score_eq1, score_eq2 = int(row["Score Eq1"]),int(row["Score Eq2"])
    week = row["week"]
    
    # Calculer la différence de score et déterminer le statut (gagné/perdu)
    diff_score = abs(score_eq1 - score_eq2)
    eq1_won = score_eq1 > score_eq2
    eq2_won = not eq1_won
    
    # Ajouter les joueurs avec leurs informations respectives
    expanded_rows.extend([
        {"Num Court": num_court, "Joueur": eq1_j1, "Coéquipier": eq1_j2, "Adversaires": f"{eq2_j1}, {eq2_j2}", 
         "Diff Score": diff_score, "Résultat": "Gagné" if eq1_won else "Perdu","week": week},
        {"Num Court": num_court, "Joueur": eq1_j2, "Coéquipier": eq1_j1, "Adversaires": f"{eq2_j1}, {eq2_j2}", 
         "Diff Score": diff_score, "Résultat": "Gagné" if eq1_won else "Perdu","week": week},
        {"Num Court": num_court, "Joueur": eq2_j1, "Coéquipier": eq2_j2, "Adversaires": f"{eq1_j1}, {eq1_j2}", 
         "Diff Score": diff_score, "Résultat": "Gagné" if eq2_won else "Perdu","week": week},
        {"Num Court": num_court, "Joueur": eq2_j2, "Coéquipier": eq2_j1, "Adversaires": f"{eq1_j1}, {eq1_j2}", 
         "Diff Score": diff_score, "Résultat": "Gagné" if eq2_won else "Perdu","week": week},
    ])

# Convertir en DataFrame
df_expanded = pd.DataFrame(expanded_rows)
df_expanded.loc[df_expanded["Résultat"] == "Perdu", "Diff Score"] *= -1




df_expanded[(df_expanded["Joueur"]=="Théophane F.") & (df_expanded["Coéquipier"]=="Clement L.")]

df_expanded.to_pickle("data_preprocess/df_resultat_expanded.pkl")

df_expanded.to_csv("data_preprocess/df_resultat_expanded.csv",index=False)
